requirements.txt
Flask==2.3.2
openai==0.27.0
python-pptx==0.6.21
gunicorn==20.1.0
colorthief
